Reduce SDR sample rate so as to not drop any samples
The callback function being called by `rtlsdr_read_async()`
blocks the radio from writing any more samples to the buffer
currently in possession by the callback.

This means I'm dropping samples while the processing loop in
rtl_cb() is running.
After dropping the sample rate from 2.4MSps to 1.2 to 0.8 to 0.4,
it finally looks like it's not dropping samples anymore.

This isn't the greatest solution...
Next steps would be to only use the callback to copy the buffer to some
other section of memory where a consumer thread can loop through
processing it constantly.